remove the indirection
