+++ src/AutoMapper/Configuration/Conventions/NameSplitMember.cs
@@ -1,3 +1,4 @@
+using AutoMapper.Internal;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
