+++ src/AutoMapper/ConstructorParameterMap.cs
@@ -1,3 +1,4 @@
+using AutoMapper.Internal;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
