+++ src/AutoMapper/Execution/DelegateFactory.cs
@@ -5,6 +5,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Linq.Expressions;
 using AutoMapper.Configuration;
+using AutoMapper.Internal;
 using AutoMapper.Mappers.Internal;
 
 namespace AutoMapper.Execution
