+++ src/AutoMapper/Internal/ExpressionFactory.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using AutoMapper.Configuration;
-using AutoMapper.QueryableExtensions;
 
 namespace AutoMapper.Internal
 {
@@ -15,14 +13,98 @@ namespace AutoMapper.Internal
     [EditorBrowsable(EditorBrowsableState.Never)]
     public static class ExpressionFactory
     {
-        public static LambdaExpression MemberAccessLambda(Type type, string memberPath) =>
-            MemberAccessLambda(ReflectionHelper.GetMemberPath(type, memberPath));
-
-        public static LambdaExpression MemberAccessLambda(params MemberInfo[] members)
+        public static Expression Chain(this IEnumerable<Expression> expressions, Expression parameter) => expressions.Aggregate(parameter,
+            (left, right) => right is LambdaExpression lambda ? lambda.ReplaceParameters(left) : right.Replace(right.GetChain().FirstOrDefault().Target, left));
+        public static LambdaExpression Lambda(this MemberInfo member) => new[] { member }.Lambda();
+        public static LambdaExpression Lambda(this IEnumerable<MemberInfo> members)
         {
-            var source = Parameter(members[0].DeclaringType, "source");
-            return Lambda(members.MemberAccesses(source), source);
+            var source = Parameter(members.First().DeclaringType, "source");
+            return Expression.Lambda(members.MemberAccesses(source), source);
         }
+
+        public static Expression MemberAccesses(this IEnumerable<MemberInfo> members, Expression obj) =>
+            members
+                .Aggregate(
+                        obj,
+                        (inner, getter) => getter is MethodInfo method ?
+                            (getter.IsStatic() ? Call(null, method, inner) : (Expression)Call(inner, method)) :
+                            MakeMemberAccess(getter.IsStatic() ? null : inner, getter));
+
+        public static IEnumerable<MemberInfo> GetMembersChain(this LambdaExpression lambda) => lambda.Body.GetMembersChain();
+
+        public static MemberInfo GetMember(this LambdaExpression lambda) =>
+            (lambda?.Body is MemberExpression memberExpression && memberExpression.Expression == lambda.Parameters[0]) ? memberExpression.Member : null;
+
+        public static IEnumerable<MemberInfo> GetMembersChain(this Expression expression) => expression.GetChain().Select(m => m.MemberInfo);
+
+        public static IEnumerable<Member> GetChain(this Expression expression)
+        {
+            return GetMembersCore().Reverse();
+            IEnumerable<Member> GetMembersCore()
+            {
+                Expression target;
+                MemberInfo memberInfo;
+                while (expression != null)
+                {
+                    switch (expression)
+                    {
+                        case MemberExpression member:
+                            target = member.Expression;
+                            memberInfo = member.Member;
+                            break;
+                        case MethodCallExpression methodCall when methodCall.Method.IsStatic:
+                            if (methodCall.Arguments.Count == 0 || !methodCall.Method.Has<ExtensionAttribute>())
+                            {
+                                yield break;
+                            }
+                            target = methodCall.Arguments[0];
+                            memberInfo = methodCall.Method;
+                            break;
+                        case MethodCallExpression methodCall:
+                            target = methodCall.Object;
+                            memberInfo = methodCall.Method;
+                            break;
+                        default:
+                            yield break;
+                    }
+                    yield return new Member(expression, memberInfo, target);
+                    expression = target;
+                }
+            }
+        }
+        public readonly struct Member
+        {
+            public Member(Expression expression, MemberInfo memberInfo, Expression target)
+            {
+                Expression = expression;
+                MemberInfo = memberInfo;
+                Target = target;
+            }
+            public Expression Expression { get; }
+            public MemberInfo MemberInfo { get; }
+            public Expression Target { get; }
+        }
+        public static IEnumerable<MemberExpression> GetMemberExpressions(this Expression expression)
+        {
+            var memberExpression = expression as MemberExpression;
+            if (memberExpression == null)
+            {
+                return Array.Empty<MemberExpression>();
+            }
+            return expression.GetChain().Select(m => m.Expression as MemberExpression).TakeWhile(m => m != null);
+        }
+        public static void EnsureMemberPath(this LambdaExpression exp, string name)
+        {
+            if (!exp.IsMemberPath())
+            {
+                throw new ArgumentOutOfRangeException(name, "Only member accesses are allowed. " + exp);
+            }
+        }
+
+        public static bool IsMemberPath(this LambdaExpression exp) => exp.Body.GetMemberExpressions().FirstOrDefault()?.Expression == exp.Parameters.First();
+
+        public static LambdaExpression MemberAccessLambda(Type type, string memberPath) =>
+            ReflectionHelper.GetMemberPath(type, memberPath).Lambda();
 
         public static Expression GetSetter(MemberExpression memberExpression)
         {
@@ -101,7 +183,7 @@ namespace AutoMapper.Internal
 
         public static Expression ToType(Expression expression, Type type) => expression.Type == type ? expression : Convert(expression, type);
 
-        public static Expression ReplaceParameters(LambdaExpression exp, params Expression[] replace)
+        public static Expression ReplaceParameters(this LambdaExpression exp, params Expression[] replace)
         {
             var replaceExp = exp.Body;
             for (var i = 0; i < Math.Min(replace.Length, exp.Parameters.Count); i++)
@@ -109,7 +191,7 @@ namespace AutoMapper.Internal
             return replaceExp;
         }
 
-        public static Expression ConvertReplaceParameters(LambdaExpression exp, params Expression[] replace)
+        public static Expression ConvertReplaceParameters(this LambdaExpression exp, params Expression[] replace)
         {
             var replaceExp = exp.Body;
             for (var i = 0; i < Math.Min(replace.Length, exp.Parameters.Count); i++)
@@ -117,11 +199,11 @@ namespace AutoMapper.Internal
             return replaceExp;
         }
 
-        public static Expression Replace(Expression exp, Expression old, Expression replace) => new ReplaceExpressionVisitor(old, replace).Visit(exp);
+        public static Expression Replace(this Expression exp, Expression old, Expression replace) => new ReplaceExpressionVisitor(old, replace).Visit(exp);
 
         public static LambdaExpression Concat(LambdaExpression expr, LambdaExpression concat) => (LambdaExpression)new ExpressionConcatVisitor(expr).Visit(concat);
 
-        public static Expression NullCheck(Expression expression, Type destinationType = null)
+        public static Expression NullCheck(this Expression expression, Type destinationType = null)
         {
             destinationType ??= expression.Type;
             var chain = expression.GetChain().ToArray();
@@ -176,7 +258,7 @@ namespace AutoMapper.Internal
             return TryFinally(body, disposeCall);
         }
 
-        public static Expression IfNullElse(Expression expression, Expression then, Expression @else = null)
+        public static Expression IfNullElse(this Expression expression, Expression then, Expression @else = null)
         {
             var nonNullElse = ToType(@else ?? Default(then.Type), then.Type);
             if(expression.Type.IsValueType && !expression.Type.IsNullableType())
@@ -264,7 +346,7 @@ namespace AutoMapper.Internal
                 return base.Visit(node);
             }
 
-            protected override Expression VisitLambda<T>(Expression<T> node) => Lambda(Visit(node.Body), node.Parameters);
+            protected override Expression VisitLambda<T>(Expression<T> node) => Expression.Lambda(Visit(node.Body), node.Parameters);
         }
     }
 }
