+++ src/AutoMapper/Mappers/ToStringDictionaryMapper.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
+using AutoMapper.Internal;
 using AutoMapper.Mappers.Internal;
 
 namespace AutoMapper.Mappers
