+++ src/AutoMapper/QueryableExtensions/Impl/QueryMapperVisitor.cs
@@ -1,3 +1,4 @@
+using AutoMapper.Internal;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
